+++ graph_attention_student/util.py
@@ -292,6 +292,14 @@ def normalize_importances_fidelity(importances: np.ndarray,
 
 def array_normalize(array: np.ndarray
                     ) -> np.ndarray:
+    """
+    Normalizes all the values of the given ``array`` into a range between 0 and 1 based on the values 
+    maximum and minimum values of the array.
+    
+    :param array: The array to be normalized
+    
+    :returns: The normalized array 
+    """
     norm = mcolors.Normalize(vmin=np.min(array), vmax=np.max(array))
     return np.vectorize(norm)(array)
 
@@ -299,6 +307,15 @@ def array_normalize(array: np.ndarray
 def binary_threshold(array: np.ndarray,
                      threshold: float,
                      ) -> np.ndarray:
+    """
+    Applies a binary threshold to the given ``array``. All values that are greater than the given 
+    ``threshold`` will be set to 1, all other values will be set to 0.
+    
+    :param array: The array to be thresholded
+    :param threshold: The threshold value
+    
+    :returns: The thresholded binary array
+    """
     binary = np.zeros_like(array)
     binary[array > threshold] = 1
     return binary