+++ HISTORY.rst
@@ -172,7 +172,7 @@ python during the forward pass and the training step implementation)
   - ``vgd_torch__megan__aqsoldb.py``
   - ``vgd_torch__megan__mutagenicity.py``
 
-0.14.1 - xx.xx.2024
+0.15.0 - 10.03.2024
 -------------------
 
 - Created a new experiment module ``vgd_torch__megan__fia_49k.py`` which trains the MEGAN model on the FIA dataset 
@@ -181,3 +181,11 @@ python during the forward pass and the training step implementation)
   loss. This is a common practice in the literature and should improve the performance of the model.
 - Added the functionality to save the MEGAN model as a PT file to the disk
 - Changed the python dependency to allow also newer versions of python
+
+0.15.1 - 22.03.2024
+-------------------
+
+. Changed the ``vgd_torch.py`` base experiment to now define the train test split with a hook because that should be 
+  more customizable in the future
+- Added the ``predict_graph`` function to the torch model base class which predicts a single graph output to be consistent 
+  with the tensorflow version